- 结构重排列
