--- src/main/utils/NumberUtils.go
@@ -1,5 +1,0 @@
-package main
-
-func isNotZero(i int) bool {
-	return i != 0
-}
