+++ src/syoka/practical/exception/ioexception.go
@@ -1,4 +1,4 @@
-package main
+package exception
 
 import (
 	"fmt"
@@ -6,13 +6,6 @@ import (
 	"math/rand"
 	"time"
 )
-
-func main() {
-	content, err := ioErr()
-	fmt.Printf("[result->(%+v),exception->(%+v)]\n", content, err)
-	result, err := algErr()
-	fmt.Printf("[result->(%+v),exception->(%+v)]\n", result, err)
-}
 
 //网络异常，编译声明异常
 func ioErr() (string, error) {
@@ -22,7 +15,6 @@ func ioErr() (string, error) {
 
 //运行时异常，手动判断？
 func algErr() (int, error) {
-
 	rand.Seed(time.Now().UnixNano())
 	total := rand.Intn(1)
 	sum := 1
