+++ src/syoka/practical/go.mod
@@ -1,0 +1,2 @@
+module syoka/practical
+go 1.23
