+++ src/syoka/practical/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"syoka/practical/utils"
+)
+
+func main() {
+	println(utils.IsEmpty("hello"))
+}
