+++ src/syoka/practical/pointer/Pointer.go
@@ -1,4 +1,4 @@
-package main
+package pointer
 
 import "fmt"
