+++ src/syoka/practical/statistic/stringConcatTest.go
@@ -1,4 +1,4 @@
-package main
+package statistic
 
 import (
 	"bytes"
