+++ src/syoka/practical/structs/Transportation.go
@@ -1,4 +1,4 @@
-package main
+package structs
 
 import "fmt"
