+++ src/syoka/practical/testing/testfunction.go
@@ -1,0 +1,8 @@
+package testing
+
+import (
+	"testing"
+)
+
+func test(t *testing.T) {
+}
