+++ src/syoka/practical/utils/stringUtil.go
@@ -1,0 +1,11 @@
+package utils
+
+import "strings"
+
+func IsEmpty(str string) bool {
+	return len(strings.ReplaceAll(str, " ", "")) != 0
+}
+
+func IsNotEmpty(str string) bool {
+	return !IsEmpty(str)
+}
